add hidden coordinate debug p1 disconnected, p4 connected on coordinate viewer dpad moves analog stick value, hold x or y to exclusively move that value hold l to move one unit per-press

+++ include/waveform.h
@@ -42,8 +42,8 @@ typedef struct MeleeCoordinates {
 	int8_t cStickX;
 	int8_t cStickY;
 
-	int stickMagnitude;
-	int cStickMagnitude;
+	double stickMagnitude;
+	double cStickMagnitude;
 
 	double stickAngle;
 	double cStickAngle;
